pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "team5 main project"
readme = "README.md"
requires-python = ">=3.12"

dependencies = [
    "django>=5.2.7,<5.3",
    "djangorestframework>=3.16.1,<3.17",
    "djangorestframework-simplejwt>=5.3.1,<5.4",
    "drf-spectacular (>=0.29.0,<0.30.0)",
    "drf-spectacular-sidecar (>=2025.10.1,<2026.0.0)",
    "python-dotenv>=1.0.1,<2.0.0",
    "django-cors-headers>=4.4.0,<5.0.0"
]

[dependency-groups]
dev = [
    "black",
    "flake8",
    "pytest",
    "pytest-django",
    "coverage",
    "python-dotenv"
]

prod = [
    "gunicorn>=21.0.0,<22.0.0",
    "psycopg2-binary>=2.9.0,<3.0.0"
]
